QuestionFactory implementation connection QuestionFactory with Controller
