+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,5 @@
+struct QuizQuestion{
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
