+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,5 @@
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
