+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct QuizStepViewModel{
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
